requirement.txt
django
paddlepaddle
paddleocr
pillow
openpyxl
djangorestframework